toby-spring/src/test/java/me/bingbingpa/vol1/ch2/step1/UserDaoTest.java
package me.bingbingpa.vol1.ch2.step1;

import me.bingbingpa.vol1.ch2.domain.User;
import org.junit.jupiter.api.Test;
import org.springframework.context.ApplicationContext;
import org.springframework.context.annotation.AnnotationConfigApplicationContext;

import java.sql.SQLException;

import static org.hamcrest.MatcherAssert.assertThat;
import static org.hamcrest.Matchers.is;

class UserDaoTest {

    @Test
    void andAndGet() throws SQLException, ClassNotFoundException {
        ApplicationContext context = new AnnotationConfigApplicationContext(UserDaoFactory.class);
        UserDao dao = context.getBean("userDao", UserDao.class);

        User user = new User();
        user.setId("bingbingpa");
        user.setName("shpark");
        user.setPassword("test");

        dao.delete();
        dao.add(user);

        User user2 = dao.get(user.getId());

        assertThat(user2.getName(), is(user.getName()));
        assertThat(user2.getPassword(), is(user.getPassword()));
    }
}